py/iteam/trae/data-assimilation-platform/algorithm_core/src/data_sources/satellite.py
# data_sources/satellite.py
# 卫星数据源处理

import numpy as np
from typing import Optional, List, Dict, Any, Tuple
import h5py
import netCDF4 as nc
import logging

from .base import DataSourceBase

logger = logging.getLogger(__name__)


class SatelliteDataSource(DataSourceBase):
    """
    卫星数据源
    处理卫星观测数据
    """
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        super().__init__(config)
        self.satellite_type = self.config.get('satellite_type', 'GOES-16')
        self.data_format = self.config.get('data_format', 'netcdf')
    
    def load_data(self, file_path: str, *args, **kwargs) -> bool:
        """
        加载卫星数据
        
        Args:
            file_path: 数据文件路径
        
        Returns:
            bool: 加载是否成功
        """
        try:
            if self.data_format == 'netcdf':
                self._load_netcdf(file_path)
            elif self.data_format == 'hdf5':
                self._load_hdf5(file_path)
            else:
                logger.error(f"不支持的数据格式: {self.data_format}")
                return False
            
            logger.info(f"成功加载卫星数据: {file_path}")
            return True
        except Exception as e:
            logger.error(f"加载卫星数据失败: {e}")
            return False
    
    def _load_netcdf(self, file_path: str):
        """
        加载NetCDF格式的卫星数据
        """
        with nc.Dataset(file_path, 'r') as ds:
            self.data = {}
            for var_name in ds.variables:
                self.data[var_name] = ds[var_name][:]
            
            # 提取元数据
            self.metadata['satellite'] = getattr(ds, 'satellite', self.satellite_type)
            self.metadata['instrument'] = getattr(ds, 'instrument', 'unknown')
            self.metadata['time'] = getattr(ds, 'time', None)
            self.metadata['spatial_resolution'] = getattr(ds, 'spatial_resolution', 'unknown')
    
    def _load_hdf5(self, file_path: str):
        """
        加载HDF5格式的卫星数据
        """
        with h5py.File(file_path, 'r') as f:
            self.data = {}
            def _extract_data(group, prefix=''):
                for key in group.keys():
                    item = group[key]
                    if isinstance(item, h5py.Group):
                        _extract_data(item, f"{prefix}{key}/")
                    else:
                        self.data[f"{prefix}{key}"] = item[:]
            
            _extract_data(f)
            
            # 提取元数据
            self.metadata['satellite'] = self.satellite_type
            self.metadata['instrument'] = 'unknown'
            self.metadata['time'] = None
            self.metadata['spatial_resolution'] = 'unknown'
    
    def process_data(self) -> Any:
        """
        处理卫星数据
        
        Returns:
            Any: 处理后的数据
        """
        if not self.validate_data():
            return None
        
        try:
            # 处理不同类型的卫星数据
            if 'brightness_temperature' in self.data:
                return self._process_thermal_data()
            elif 'reflectance' in self.data:
                return self._process_reflectance_data()
            else:
                logger.warning("未知的卫星数据类型")
                return self.data
        except Exception as e:
            logger.error(f"处理卫星数据失败: {e}")
            return None
    
    def _process_thermal_data(self) -> Dict[str, Any]:
        """
        处理热红外数据
        """
        # 转换亮温数据为温度
        brightness_temp = self.data['brightness_temperature']
        # 这里可以添加更复杂的处理逻辑
        return {'temperature': brightness_temp}
    
    def _process_reflectance_data(self) -> Dict[str, Any]:
        """
        处理反射率数据
        """
        # 处理反射率数据
        reflectance = self.data['reflectance']
        # 这里可以添加更复杂的处理逻辑
        return {'reflectance': reflectance}
    
    def get_observations(self) -> Tuple[List[float], List[Tuple[float, float, float]], List[float]]:
        """
        获取观测数据
        
        Returns:
            Tuple[观测值列表, 观测位置列表, 观测误差列表]
        """
        if not self.validate_data():
            return [], [], []
        
        try:
            # 提取观测值
            if 'brightness_temperature' in self.data:
                obs_values = self.data['brightness_temperature'].flatten().tolist()
            elif 'reflectance' in self.data:
                obs_values = self.data['reflectance'].flatten().tolist()
            else:
                logger.warning("无法提取观测值")
                return [], [], []
            
            # 生成观测位置 (这里需要根据实际数据结构调整)
            obs_locations = []
            if 'lat' in self.data and 'lon' in self.data:
                lat = self.data['lat']
                lon = self.data['lon']
                for i in range(lat.shape[0]):
                    for j in range(lat.shape[1]):
                        obs_locations.append((lon[i, j], lat[i, j], 0.0))  # 假设高度为0
            else:
                # 生成默认位置
                for i in range(len(obs_values)):
                    obs_locations.append((0.0, 0.0, 0.0))
            
            # 生成观测误差
            obs_errors = [0.5] * len(obs_values)  # 默认误差
            
            return obs_values, obs_locations, obs_errors
        except Exception as e:
            logger.error(f"获取观测数据失败: {e}")
            return [], [], []
