py/iteam/trae/data-assimilation-platform/algorithm_core/pyproject.toml
[project]
name = "bayesian_assimilation"
version = "1.0.0"
description = "贝叶斯同化系统核心算法"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "numba>=0.57.0",
    "psutil>=5.9.0",
    "matplotlib>=3.7.0",  # 用于可视化
    "pytest>=7.4.0",      # 用于测试
    "pyyaml>=6.0",        # 用于配置文件
    "netcdf4>=1.6.0",     # 可选：用于NetCDF数据
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x>=12.0.0",  # 根据您的CUDA版本选择
    "cupyx>=0.4.0",
]

dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourname/data_assimilation_platform"
Documentation = "https://data-assimilation-platform.readthedocs.io"
Repository = "https://github.com/yourname/data_assimilation_platform"
Issues = "https://github.com/yourname/data_assimilation_platform/issues"